+++ src/Resources/app/administration/src/main.js
@@ -7,8 +7,10 @@ import './component/orders-unprocessed';
 import './component/orders-paid';
 import './component/orders-refunded';
 import './component/orders-canceled';
+import './component/config-tabs';
 
 import './extension/sw-order/view/sw-order-detail-base';
+import './extension/sw-settings-payment/page/sw-settings-payment-detail';
 
 import localeDE from '../../../snippet/storefront/worldline.de-DE.json';
 import localeEN from '../../../snippet/storefront/worldline.en-GB.json';