+++ src/Resources/app/administration/src/service/transactionsControlService.js
@@ -67,6 +67,25 @@ class transactionsControlService extends ApiService {
                 return ApiService.handleResponse(response);
             });
     }
+
+    getOneyPaymentConfig(values) {
+        const headers = this.getBasicHeaders({});
+        return this.httpClient
+            .post(`_action/${this.getApiBasePath()}/getOneyPaymentOption`, values,{
+                headers
+            })
+            .then((response) => {
+                return ApiService.handleResponse(response);
+            });
+    }
+
+    setOneyPaymentConfig(values) {
+        const headers = this.getBasicHeaders({});
+        return this.httpClient
+            .post(`_action/${this.getApiBasePath()}/setOneyPaymentOption`, values,{
+                headers
+            });
+    }
 }
 
 Application.addServiceProvider('transactionsControl', (container) => {